componentの切り出し
